Explicitly define x86_64 FPRegSet struct rather than using libc
To support cleanly support multiple architectures, we need to be able to build a binary for one architecture that manipulates snapshots and snaps for another architecture. This is not possible if we're using data structures defined by libc. So we must define the structure of FPRegSet ourselves.

This struct is designed to be compatible with fxsave64, as is the libc version, so they are quite similar. The main differences are that we used the names from the AMD and Intel manuals, which are slightly different from the libc names. We also represented the FP data registers as opaque 128-bit ints rather than structs with subfields.

PiperOrigin-RevId: 474358121