Change serialization format for x86_64 registers
This change makes x86_64 serialize its registers similar to aarch64:
prepending a header to the raw bytes.

The old serialization format relied on arch-specific system headers which
meant that x86_64 serialization could only be performed on x86_64 hosts.
This made testing more complicated and prevented tests in the common/
directory from being enabled on aarch64. Running these tests on aarch64
was not strictly necessary, but figuring out which tests should and should
not run on aarch64 was becoming complicated. This should make it simple.

The old serialization format can still be loaded, newly created Snapshots
will use the new format. The old serialization format did not have any
sort of header, so we identify the old data by its size.

PiperOrigin-RevId: 483480215