Fix PartitionEvenly when the input size is less than the partition count
Currently PartitionEvenly will index the input vector with an out-of-bounds
index when it creates empty partitions. This could be fixed one of two ways.
1) Elide empty partitions from the output.
2) Create the empty partitions safely.

All of the clients of PartitionEvenly appear to assume the number of
partitions appears to be exactly the number requested, so option 2 was
selected. For consistency, the special case where a zero-length input
would result in zero partitions was eliminated - none of the clients
expected this case.

PiperOrigin-RevId: 484304320