Pre-compile test snapshots
The current approach of scraping instructions out of inline ASM does not
work as well when we need to support multiple architectures. Instead we
generate the compiled instructions for the test snapshots using a script.

It may be possible to integrate the invocation of this script into the build
system, but invoking compilers and tools can be fiddly so for now we're
checking in the generated source.

PiperOrigin-RevId: 484657904