Eliminate InitTestSnapshotRegs dependency on host
Explicitly specify register values rather than defaulting to whatever the
current value on the host is. This makes the code deterministic and also
allows it to run on non-x86_64 hosts.

Also fix similar code in raw_insns_util.cc to initialize mxcsr_mask. An
uninitialized value would be interpreted as 0xffbf, instead explicitly
initialize it to 0xffff rather than rely on implicit re-interpretation.

PiperOrigin-RevId: 485383043