Stop inferring Arch parameter of GRegSet and FPRegSet
Previously, an unparameterized register set was inferred to match the host
arch. This was done to reduce the diffs caused by the initial host/target
split. Enough refactoring has occurred that this default is no longer
needed, and making the parameterization explicit makes it easier to see
where the portability problems are.

PiperOrigin-RevId: 485999823