Allow make Arch type param available on ucontext types
Generally this shouldn't be needed, but occasionally it is necessary to propagate type parameters to functions that do not take a parameterized
type, for example: foo<ARCH_OF(uctx)>(). In many cases this is
unnecessary because the enclosing function is itself parameterized with
Arch, but there are corner cases where the function is not parameterized
and extracting the Arch param from another variable is cleaner than
repeating the Arch type multiple times and risking inconsistency.

Also add a generic version of ucontext_test.cc and do some cleanups.

PiperOrigin-RevId: 486176282