Move RegisterState structures out of Snap structure
Currently Corpus relocation dirties the pages containing Snaps, which causes the
RegisterState structures (which are never relocated and are a majority of the
Corpus bytes) to be copied into dirty pages. Moving the registers out of line from
the Snap structure makes ~17% of the memory in the runner clean shared pages and
saves ~7% memory when running the orchestrator. The only downside is adding two
pointers increases the compressed corpus size by ~1%.

This also sets up for supporting multiple architectures.

PiperOrigin-RevId: 487660047