Search for test Snaps by name rather than assuming an index
Previously we assumed that the test Snap corpuses contained all the Snaps in
a corresponding enum and they were in the same order as the enum. The Snap could be found at the index corresponding to the value of the enum.

Now we map the enum to a name, and then do a linear search of the corpus for
a Snap with an ID matching that name. This makes it easier to support multiple architectures because not all tests will be supported on all architectures.

PiperOrigin-RevId: 492050518