Filter for collisions between code addresses and the exit sequence early
The exit sequence address (0xABCD0000) unfortunately exists in the range of addresses that instruction sequences can be placed (0x30000000-0xB0000000). This means that the filtering or the making process will reject instruction sequences at a fairly low rate (1/512k). To assist analysis of the fuzzing pipeline, reject these collisions as early as possible.

PiperOrigin-RevId: 500829332