Filter out aarch64 PAC instructions
These instructions depend random values set by the kernel. We might be able to control
theses values with ptrace, but for the moment these instructions are effectively
non-deterministic.

It also appears that the version of QEMU we are using does not support PAC and interprets
these instructions as normal loads and stores, letting the throught and causing faults in
the maker.

PiperOrigin-RevId: 501702973