Fix SessionSummary.max_rss_kb contents.
The value reported by getrusage(RUSAGE_CHILDREN) is bogus.
In almost all cases it will be the RSS of the orchestrator process itself
and not the largest runner. This is because "maxrss is effectively
inherited" over fork/exec (https://jkz.wtf/random-linux-oddity-1-ru_maxrss)
and the fact that the orchestrator loads all corpus shards into memory
first raising the maxrss value from the start.
Replace getrusage() with periodically sampling the RSS from /proc/

PiperOrigin-RevId: 505797889