GetCPUIdNoSyscalls: fall back to reporting CPU affinity
In some cases (aarch64, primarily) it appears we may not be able to get the current CPU number without making a syscall. There does not appear to be a register / instruction like RDTSCP on x86. To provide a reasonable guess about the CPU number, we record the CPU affinity when it is set and assume that is the CPU we are running on.

PiperOrigin-RevId: 513021490