Add real-time timeout for UnicornTracer
Bugs in QMEU can cause Unicorn to hang. For example, on aarch64 in some cases the ld4 instruction can hang. In some cases it can also crash. It's likely a codegen bug that we can fix by pulling in a new version of Unicorn, but this is quick first aid to prevent fuzzing and analysis pipelines from getting bogged down.

PiperOrigin-RevId: 553538293