Ensure InvalidInstructionID() is within [0,NumInstructionIDs())
Previously InvalidInstructionID() was UINT32_MAX, which meant it need to be special cased. Now that it live in the same range as the other instruction IDs, invalid instructions can be analyzed the same way as valid instructions.

As it happened, all the disassemblers we use already reserved zero as the ID for an invalid instruction, so this change was not complicated.

PiperOrigin-RevId: 555295796