Simplify OwnedFileDescriptor
Previously OwnedFileDescriptor was a unique_ptr<int*> with a special destructor. This meant using an owned file descriptor required explicitly allocating memory and derefing to access the actual FD.

Refactor it into a flat no-copy wrapping type to remove any suggestion at the use sites that this is a pointer. This requires writing a little more code than before, but it gives us control over the API.

PiperOrigin-RevId: 570471896