Make proxies compatible with Unicorn v2
The default CPU settings for v2 are lower than we'd like and we'd lose a bit of coverage.

Due to Silifuzz having two slightly different builds and the complexities or upgrading third-party software, the v2 calls are temporally gated by the preprocessor so that the source stays compatible with v1.

PiperOrigin-RevId: 582379927