Fix off-by-one error in InstructionData::Copy
The method was checking that the incoming data was smaller than the internal buffer and this mean that last byte of the buffer could not be used. In practice this would not be an issue because the buffer size was rounded up to a power of two, so the last byte would never be used anyways.

Also micro optimize Program::ToBytes by reserving space in the output vector.

PiperOrigin-RevId: 587777216