Configure unicorn_aarch64 proxy to run as A72
This is probably a bit older than most users would want, but the only other alternative that's immediately available is a bit too new and accepts instructions that will not run on most hardware that is currently available. Longer term we need to better support a wider variety of microarchitechtures.

trace_tool was not affected by these changes and it still should be able to trace all the instructions Unicorn can trace.

PiperOrigin-RevId: 599000955