Mask the pstate bits provided by the Linux signal handler on aarch64
The Linux signal handler provides more bits from pstate than Silifuzz can capture on the normal SaveUContext path. We mask the value provided by the signal handler to match the value we'd obtain using the other path.

This skew did not cause false positives because a signal is always an error. It did, however, cause noise in the register diffs when triaging errors.

PiperOrigin-RevId: 601171658