Abstract out arch-dependant instruction sizes in Silifuzz mutator
As an issue of correctness, when the mutator fails to decompile the instruction the PC should always be incremented by 4 on aarch64, instead of 1 on X86_64.

Fixing this issue also opened a few opportunities to make the code  a little more efficient on aarch64, due to the fixed instruction size.

PiperOrigin-RevId: 603812768