#silifuzz Map data pages while making snapshot.
This enables the runner option --max_pages_to_add.  Setting it to a non-zero value with --make allows the runner to continue making when a SEGV fault due to a missing page is received.

This only changes the runner. Related changes in the runner driver and the snap maker will appear in subsequent changes.

PiperOrigin-RevId: 605065089