#silifuzz Use MAP_FIXED_NOREPLACE flag for new mapping during snapshot making.
This adds an additional mmap flag to prevent the runner from mmapping an existing page during making. This happens when the runner tries to recover from a page crossing memory fault of an emulated instruction. On the x86, User Mode Instruction Prevention (UMIP) security extension causes the Linux kernel to emulate some instructions. When the first page of a page crossing access is present but the second page is absent, the kernel reports an address in the first page in siginfo accompanying a SEGV fault.

PiperOrigin-RevId: 609432576