#silifuzz Restore UContext from a UContextView.
This changes how we restore context. Previously RestoreUContext*() reads from a UContext struct. This is now changed to read from a const reference to a UContextView object, which contain separate pointers to FPRegs and GPRegs. Doing so allows us to restore from registers that are not necessarily stored adjacently in memory. This will enable us to de-duplicate the floating point initial states for snaps, which are all there same currently.

PiperOrigin-RevId: 617622073