#silifuzz Split snap register memory checksum by register set.
This changes how we compute snap register memory checksum for verifying data integrity of a snap when a snap does not end as expected. Previous we use a single checksum for both GPRs and FPRs contained in a single block of memory. Since introduction of UContextView, the registers may not be physically together. This change breaks the original single checksum into two parts, one for each register set.

PiperOrigin-RevId: 618317471