#silifuzz Improve CRC32C performance.
We already use hardware CRC instructions on both x86 and aarch64 but we suffer a 3-cycle latency on x86 and possibly something similar on aarch64. This is mostly an issue on x86 as we do not do SIMD checksumming on aarch64 currently. To hide CRC instruction latency, we split a block of input into 3 equally-sized parts and compute 3 CRC values in parallel. The partial results are then combined into a single value. This is profitable only for big enough block sizes as combining CRC values is expensive.

To speed up CRC combination, we use look up tables. These tables are read only and in generated source code.

PiperOrigin-RevId: 625030507