Pull out the Centipede defs library into a new common package.
With more and more cross-dependencies between FuzzTest and Centipede libraries,
it's starting to make sense to have a package for libraries shared between both
modules.

This is the first step in that direction, with the aim to extract remote_file
and blob_file so that they can be used in FuzzTest.

PiperOrigin-RevId: 646259827