hashtest: ensure test log gets periodically flushed
For performance, it's better to send "\n" than std::endl to std::cout and avoid unnecessary flushes. This causes data loss in cases where hashtest provokes a kernel panic, however, and and stdout was being redirected to a file. To support this case it's better to use std::endl. The test runner is not outputting that much data, anyways.

PiperOrigin-RevId: 675218676