Prevent random code addresses from colliding with exit page
Due to historical baggage, the exit page exists in the default address range that tests are mapped. In rare cases (1 out of 2^19) the address generated for test code will overlap the address range page, but this won't be noticed until later in the pipeline. Typically this would result in a soft error where a small percentage of tests were lost in the making pipeline. hashtest_generator_integration_test expected that all its snapshots would be valid, however, so the rare address collision caused the test to be flaky.

Instead of making the test more accepting of this error, it was simpler to fix the root cause. Because we rely on the code => address mapping to be deterministic and stable, so changing that function radically would be complicated. However - because the case where it collided with the exit page was already broken, it's possible to change how address generation works in that specific case. In the case of collision, we now try to re-randomized the address.

PiperOrigin-RevId: 703542831