Make runner location an explicit parameter of MakingConfig::Default
Previously, every time a making config was created the file system would be probed to find the runner even if the path to the runner would be supplied later. Making the runner path a parameter to config creation allows tighter control of when the probing does and does not happen. Effectively the code is better structured for dependency injection.

PiperOrigin-RevId: 704810151