Add a member to store raw instruction in DecodedInsn
Xed's API to get the raw instruction bytes `xed_decoded_inst_get_byte()` seems to rely on the original data buffer that was passed to `xed_decode()` to return the correct values. Added a field that saves a copy of raw insn bytes for them to be reliably accessible, even after the original data buffer is out of scope.

PiperOrigin-RevId: 780311846