Restrict the snapshot to the expected end state only in snap_tool.
RunnerDriver expects only a single endstate and assumes that the first end state
is the intended one.  This can cause snapshots with multiple end states to fail
when they actually should be passing.  Restricting the set of end states to the
platform value specified in the command line (or the current platform if unspecified)

PiperOrigin-RevId: 842780371